validate: Extract shared test helpers to validate_test.go
Move shared constants, fixtures (testSystem, testK8s, testOcp), and
helper functions (testCommand, checkReport, checkStep, etc.) from
command_test.go to validate_test.go. This prepares for splitting the
command flow tests into clusters_test.go and application_test.go.
